Use double precision in fusion filters and PID

+++ IMU/main/Complementary_Filter.h
@@ -21,10 +21,10 @@ class Complementary_Filter : public Fusion_Filter
 private:
     IMU::Euler_Angles euler;
     timeval last_time;
-    const float COMPLEMENTARY_COEFFICIENT;
+    const double COMPLEMENTARY_COEFFICIENT;
 
 public:
-    Complementary_Filter(float coeff);
+    Complementary_Filter(double coeff);
     ~Complementary_Filter() override = default;
 
     void Compute(IMU::Sensor_Data accel, IMU::Sensor_Data gyro, IMU::Euler_Angles &euler) override;
